download a github release asset to the stow R user directory

Use stow to abstract away the process of downloading a file or a GitHub release asset to a user's data directory, only downloading files that have not already been downloaded.

Usage

stow_gh_release(owner, repo, dpkg, overwrite = FALSE)

stow(uri, overwrite = FALSE)

stow_url(url, overwrite = FALSE)

Arguments

owner

string of repo owner

repo

string of repo name

dpkg

string of gh release tag (will be the same as the filename without the .parquet extension)

overwrite

logical; re-download the remote file even though a local file with the same name exists?

uri

character string universal resource identifier; currently, must begin with http://, https://, or gh://

url

a URL string starting with http:// or https://

Value

path to the stowed file or url to github release

Details

Supported URI prefixes include:

https://, http://: download from a file

gh://: download a github release asset, formatted as gh://owner/repo/name

Stow downloads files to the users data directory; see ?tools::R_user_dir. Specify an alternative download location by setting the R_USER_DATA_DIR environment variable. The stow cache works by name only; that is, if a file with the same URI has already been downloaded once, it will not be re-downloaded again (unless overwrite = TRUE).

Examples

# get by using URL
stow("https://github.com/geomarker-io/appc/releases/download/v0.1.0/nei_2020.rds")
#> /home/runner/.local/share/R/stow/nei_2020.rds

# will be faster (even in later R sessions) next time
stow("https://github.com/geomarker-io/appc/releases/download/v0.1.0/nei_2020.rds") |>
  readRDS()
#> # A tibble: 369,063 × 4
#>    pollutant_code total_emissions emissions_uom s2            
#>    <chr>                    <dbl> <chr>         <s2_cell>     
#>  1 EC                  0.00137    TON           -3.419770e-306
#>  2 NO3                 0          TON           -3.419770e-306
#>  3 OC                  0.0317     TON           -3.419770e-306
#>  4 PM25-PRI            0.782      TON           -3.419770e-306
#>  5 PMFINE              0.746      TON           -3.419770e-306
#>  6 SO4                 0.00278    TON           -3.419770e-306
#>  7 EC                  0.0000182  TON           -3.355024e-306
#>  8 NO3                 0.00000638 TON           -3.355024e-306
#>  9 OC                  0.000129   TON           -3.355024e-306
#> 10 PM25-PRI            0.0007     TON           -3.355024e-306
#> # ℹ 369,053 more rows

# get a data package parquet file created with dpkg_gh_release()
stow("gh://cole-brokamp/dpkg/cagis_parcels-v0.1.0") |>
  arrow::read_parquet()
#> Error in httr2::req_perform(httr2::req_headers(httr2::request(glue::glue("https://api.github.com/repos/{owner}/{repo}/releases/tags/{dpkg}")),     Accept = "application/vnd.github+json", Authorization = glue::glue("Bearer {get_gh_token()}"),     `X-GitHub-Api-Version` = "2022-11-28", .redact = "Authorization")): HTTP 404 Not Found.